TRX: Minor fixes, especially handle TOA of RACH correctly

+++ src/osmo-bts-trx/scheduler.h
@@ -16,7 +16,7 @@ int trx_sched_tch_req(struct trx_l1h *l1h, struct osmo_phsap_prim *l1sap);
 int trx_sched_clock(uint32_t fn);
 
 int trx_sched_ul_burst(struct trx_l1h *l1h, uint8_t tn, uint32_t fn,
-        sbit_t *bits, int8_t rssi, int16_t toa);
+        sbit_t *bits, int8_t rssi, float toa);
 
 /* set multiframe scheduler to given pchan */
 int trx_sched_set_pchan(struct trx_l1h *l1h, uint8_t tn,
